Make type expression package naming consistent
