+++ src/test/kotlin/sigma/syntax/typeExpressions/OrderedTupleTypeConstructorTermTests.kt
@@ -1,4 +1,4 @@
-package sigma.syntax.type_expressions
+package sigma.syntax.typeExpressions
 
 import sigma.evaluation.values.Symbol
 import sigma.semantics.StaticScope
@@ -6,9 +6,6 @@ import sigma.semantics.types.BoolType
 import sigma.semantics.types.IntCollectiveType
 import sigma.semantics.types.OrderedTupleType
 import sigma.syntax.SourceLocation
-import sigma.syntax.typeExpressions.OrderedTupleTypeConstructorTerm
-import sigma.syntax.typeExpressions.TypeExpressionTerm
-import sigma.syntax.typeExpressions.TypeReferenceTerm
 import utils.FakeDeclarationBlock
 import utils.FakeTypeEntityDefinition
 import kotlin.test.Test
