+++ src/test/kotlin/sigma/syntax/typeExpressions/TypeReferenceTermTests.kt
@@ -1,32 +1,25 @@
-package sigma.syntax.type_expressions
+package sigma.syntax.typeExpressions
 
-import sigma.evaluation.values.Symbol
-import sigma.semantics.types.ArrayType
-import sigma.semantics.types.BoolType
 import sigma.syntax.SourceLocation
-import sigma.syntax.typeExpressions.ArrayTypeConstructorTerm
-import sigma.syntax.typeExpressions.TypeExpressionTerm
-import sigma.syntax.typeExpressions.TypeReferenceTerm
+import sigma.semantics.types.BoolType
+import sigma.evaluation.values.Symbol
 import utils.FakeDeclarationBlock
 import utils.FakeTypeEntityDefinition
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
-class ArrayTypeConstructorTermTests {
+class TypeReferenceTermTests {
     object ParsingTests {
         @Test
         fun test() {
             val expression = TypeExpressionTerm.parse(
-                source = "[A*]",
+                source = "Foo",
             )
 
             assertEquals(
-                expected = ArrayTypeConstructorTerm(
+                expected = TypeReferenceTerm(
                     location = SourceLocation(lineIndex = 1, columnIndex = 0),
-                    elementType = TypeReferenceTerm(
-                        location = SourceLocation(lineIndex = 1, columnIndex = 1),
-                        referee = Symbol.of("A"),
-                    ),
+                    referee = Symbol.of("Foo"),
                 ),
                 actual = expression,
             )
@@ -37,20 +30,18 @@ class ArrayTypeConstructorTermTests {
         @Test
         fun test() {
             val type = TypeExpressionTerm.parse(
-                source = "[A*]",
+                source = "Foo",
             ).evaluate(
                 declarationScope = FakeDeclarationBlock.of(
                     FakeTypeEntityDefinition(
-                        name = Symbol.of("A"),
+                        name = Symbol.of("Foo"),
                         definedTypeEntity = BoolType,
                     ),
                 ),
             )
 
             assertEquals(
-                expected = ArrayType(
-                    elementType = BoolType,
-                ),
+                expected = BoolType,
                 actual = type,
             )
         }
